fix(frontend): safe cancelChatStream ack and stale FAILED stream recovery
- Fix cancelChatStream ack handler: response is now optional, promise
  resolves exactly once, only explicit success:true means cancellation
  succeeded (SocketContext.tsx)
- Fix stale stream finalization: FAILED/CANCELLED subtasks now map to
  'error' message status instead of always 'completed' (TaskStateMachine.ts)
- Split recovery tests into dedicated TaskStateMachine.recovery.test.ts
  with beforeEach/afterEach console spy isolation
- Add test for FAILED stale subtask recovery
- Add targeted JSDoc for serverConfirmedNoStream, stale terminal status
  mapping, and cancelChatStream ack semantics
